[ios] Update strings for SetUpList Notification item
This change updates the string used in the SetUpList for the description
of the Notification item. In the MagicStack, a shorter string will be
used. And on the See More screen, a longer string will be used if Price
Tracking is allowed (i.e. in the US).

Fixed: 395921277
Change-Id: Ie990b73a1f13b3c82fc648381771d33054d1ea3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6478711
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1449960}