LaunchHandlerAPI: Add file handling implementation
If the web app was launched by opening one or multiple files, the list
of file URIs will be received in the incoming intent. This CL implements
passing the list of file URIs from the intent to the launch queue, so
the web app can read and write from the files.

Bug: 409247581
Bug: 369797848
Change-Id: I0199ef7e89a60b6a708733ff16627d2b04acd9c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6461069
Reviewed-by: Peter Conn <peconn@chromium.org>
Reviewed-by: Vlad Korotkevich <vkorotkevich@google.com>
Commit-Queue: Oleksii Tkachenko <tkachenkoo@google.com>
Cr-Commit-Position: refs/heads/main@{#1449966}