Add a "heavily redacted" netlog capture mode
This CL adds a new netlog capture mode, "heavily redacted", which is
designed to provide stronger privacy guarantees than even the default
mode. It does this by enforcing an allowlist of param names that are
known to be safe. The allowlist is managed centrally to reduce the
likelihood of accidental leakage.

The goal is to make it safer to collect and share netlogs that are
captured from real devices used by real end users running real
production code. While the resulting netlogs of course contain much less
information that in default mode, the data that remains can still be
quite useful, especially when gathered into a Perfetto trace and
time-correlated with other system events and other tracing sources. For
example, just being able to see retries, even if we don't know the
contents or the destination of the request being retried, can already
provide useful insights.

The allowlist in this CL is just meant as a rough initial starting
point; more entries are expected to be added in the future.

Bug: 410018349, b:400686524
Change-Id: If1a586898a4ebb3007fc0601f94d959eadf6c736
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6455065
Reviewed-by: Mike West <mkwst@chromium.org>
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Ashley Newson <ashleynewson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1449972}