Make Cronet send netlogs to Perfetto
This CL makes Cronet send netlogs to TraceNetLogObserver, which is a
bridge between netlogs and the Chromium trace framework, i.e. Perfetto.

Example trace:
https://ui.perfetto.dev/#!/?s=15c5bd3b749dcb42ad82ae2f1a183b4598a700d3
(Look under "Global Legacy Events")

This happens automatically as soon as there is an active Perfetto
trace session that has the "netlog" track event category enabled. This
makes netlog functionality much easier to use than the traditional way
to enable netlogs in Cronet, which required individual apps to call a
specific Cronet API and manage netlog files themselves.

Note that TraceNetLogObserver is smart and only enables netlog when it
detects an active trace session; if Cronet is not being actively traced,
the netlog is not enabled and there is no performance hit.

This CL does not include any mechanism to change the capture mode; for
now, we err on the safe side and always log in "heavily redacted" mode
to sidestep any privacy implications. We shall revisit this later.

Bug: b:410018349
Change-Id: Iab4ee789dec29ae91539c13933c0c9987752f568
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6460222
Reviewed-by: mmenke <mmenke@chromium.org>
Reviewed-by: Stefano Duo <stefanoduo@google.com>
Commit-Queue: Etienne Dechamps <edechamps@google.com>
Cr-Commit-Position: refs/heads/main@{#1449988}