Support /sorry error page
Glic may redirect to an error page, or navigate to an error
page that needs to be shown to the user. We identify this page
by looking at the URL path. If detected, we resize the window
and allow the guest page to be shown
After navigating back, the main application can load and
initialize successfully.

Bug: b:411722775
Change-Id: Id6ad898e8b1bfc48b333ddaa65349faca6ae0c72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6476494
Reviewed-by: Andrea Gomez <andreaxg@google.com>
Auto-Submit: Dan Harrington <harringtond@chromium.org>
Reviewed-by: Carlos Knippschild <carlosk@chromium.org>
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Andrea Gomez <andreaxg@google.com>
Cr-Commit-Position: refs/heads/main@{#1450090}