[ChromeVox] Change panel type from kComponent to kExtensionPopup
The class is only used to initiate the ChromeVox panel.
The mojom::ViewType::kComponent constant does not have an analogous
type in a manifest V3 chrome extension.

The KComponent type will cause an error when running the MV3 extension,
a NOTREACHED in:
  RuntimeGetContextsFunction::GetFrameContexts()

kPopup has an analogous type in MV3.

See:
  extensions/common/mojom/view_type.mojom

Change-Id: If9732f8b4f47fcb268c0853b5b11c1d400713a69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6476114
Reviewed-by: Akihiro Ota <akihiroota@chromium.org>
Commit-Queue: Valerie Young <spectranaut@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1450194}