[Messaging] Open previous URL when instant message Reopen button clicked
When a tab is updated, on android/iOS (not desktop) an instant message
is shown saying "Tab X was updated" with a button to reopen the
URL. This URL and the title shown in the message should both be
the previous URL / title instead of being the latest one. This
CL fixes that.
1. TabGroupChangeNotifier was improved to have both the before and
after tabs in the notification to the MessagingBackendService.
2. The service extracts the previous URL and adds it to the
MessageAttribution as a new field.
3. The Android code was updated to use the previous URL for tab update
message. Note, for tab removal we still use the latest URL.
4. Also note that the tab title being shown in the message "Tab X
was updated" is already correct since the time it was captured the
title of the tab was still old. We should probably rename this
field to be previous URL to be more explicit.

Bug: 409852510
Fixed: 409852510
Change-Id: I43315337751082f60af34e6dad2d3de4e0a7fbfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6477407
Reviewed-by: Yuheng Huang <yuhengh@chromium.org>
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450218}