Allow devil_util to read command-line arguments from file
If devil_util receives a command-line argument @/path/to/file,
devil_util will read the contents of that response file and interpret
it as command-line arguments. This avoids "argument list too long".

If the response file ends in .zst, devil_util will read the content of
the file, decompress it via zst, and work with decompressed content.

The response file can contain multiple command-line arguments.
Each line in the response file is treated as a separate argument.

Also add a new command called "compress", which compress the given
content via zst and save the compressed content to the specified file.

Bug: 397459590
Change-Id: I021f17df3b5806f0a646037c1cb05a349b734fbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6480354
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Commit-Queue: Martin Kong <martinkong@google.com>
Cr-Commit-Position: refs/heads/main@{#1450227}