[Recent Activity Android] Bottom sheet with custom life cycle
This CL changes recent activity bottom sheet to use a custom
life cycle. This fixes the issue where a navigation in the current
tab (in background) due to sync automatically closes the bottom
sheet. Since we are using a custom life cycle, we also start observing
TabGroupSyncService for group removals so that we can close the bottom
sheet when the group is gone.

Also added a util method in TabGroupSyncUtils to find the tab group
associated with a collaboration ID which is required at a lot of places.

Fixed: 403340830
Change-Id: I38a50cd84d5f48389b3bc2678584f56802028a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6477470
Commit-Queue: Shakti Sahu <shaktisahu@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450238}