Plumb enforce_anchor_expiry and enforce_anchor_constraints to BoringSSL
https://chromium-review.googlesource.com/c/chromium/src/+/6440022
added constraints to the Chrome Root Store proto which were intended
to be used for preloading intermediates in the Chrome Root Store. If
we ship intermediates in the Chrome Root Store and treat them as trust
anchors, we need to enforce expiry and X.509 constraints on them just
as if they weren't preloaded, and these flags are intended to enable
that behavior.

This CL plumbs these flags from the root store data into
TrustStoreChrome into BoringSSL, so that BoringSSL enforces them
during path-building and certificate verification. BoringSSL already
had this functionality so this CL is just a matter of configuring
BoringSSL to enforce these fields on the certificates marked as such
in the Chrome Root Store.

To do so, we modify the root store generation tool (for the
compiled-in root store) and the code to load a root store update from
a proto from component update. In both these code paths, we pass the
flags through when creating a ChromeRootStoreData. TrustStoreChrome
then applies these flags to the anchors it creates in its underlying
bssl::TrustStoreInMemory.

Like other constraints, these new flags are only enforced in Chrome's
TLS trust store, not the EUTL trust store.

Bug: 396479339
Change-Id: Iac48e319acf1eb643b882fa055188c3fbda6ef03
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6477150
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: Emily Stark <estark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450242}