[Masonry] Fix bugs with `item-direction: row`
1. In `BuildGridAxisTracks`, the GridTrackSizingCollection constructor
was being used incorrectly, in the sense that the `track_direction`
parameter was being used for `must_create_baselines`, leaving
`track_direction` to always fallback to its default value.

2. In `MasonryLayoutAlgorithm::CreateConstraintSpaceForMeasure`, the
inline size needed to be set to `kIndefinite` so that we don't fail the
DCHECK in length_utils.cc:406 which ensures that the inline size is
indefinite so that text can flow when checking the size.

Note that row is still not entirely functional after this CL, but other
bugs will be addressed in other CLs.

Bug: 343257585
Change-Id: I5ca1782e234c01d2198d3fb3917f6c0d75e7f562
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6476517
Reviewed-by: Kurt Catti-Schmidt <kschmi@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ethan Jimenez <ethavar@microsoft.com>
Commit-Queue: Celeste Pan <celestepan@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1450247}