Apply margin to snackbar view to avoid clipping issues
This change fixes a bug where strange clipping happens to the snackbar view during its dismissal animation. This change fixes this clipping by extending the snackbar view itself into the bottom inset by applying a bottom margin.

Note: disabling clipping in XML doesn't seem to fix this issue, and neither does extending the snackbar's parent views - it has to be applied specifically to the snackbar.

This fix is only relevant when edge-to-edge is enabled, otherwise the snackbar is intended to dismiss "under" the navigation bar. The issue is specifically when the snackbar appears to dismiss "under" a transparent navigation bar in edge-to-edge mode.

Before:
https://screencast.googleplex.com/cast/NDg1NTMyNDQyNjEwODkyOHxhZDRkZGFhOC1mYQ
https://screencast.googleplex.com/cast/NjQ4MjUzMzU4NjgyOTMxMnw4MDQ2ODA5Ni1mMg

After:
https://screencast.googleplex.com/cast/NTAxNTg4MTcyNDcyMzIwMHwxMTRiMzc4Yy1mNw
https://screencast.googleplex.com/cast/NTk4NzAwNTE3MDE4ODI4OHxiYmFlYWE3Ny02Mg

E2E disabled:
https://screencast.googleplex.com/cast/NTgzODkzOTU5NTIxMDc1Mnw2MmRiZjg0Yy0zYQ
https://screencast.googleplex.com/cast/NjA5OTc4MjA4NTkwMjMzNnw5ZDI4NDJjOC05Mg

Bug: 385200055
Change-Id: I1e1ce2c80ef0679892266743dfad0338a0f57fc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6481293
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Charles Hager <clhager@google.com>
Cr-Commit-Position: refs/heads/main@{#1450291}