Refactor HttpTransactionFactory creation to use a wrapper callback
This change refactors the mechanism for customizing the
HttpTransactionFactory in URLRequestContextBuilder.

The previous `SetCreateHttpTransactionFactoryCallback` required the
callback to construct the entire transaction factory stack, given only
the HttpNetworkSession.

The new `SetHttpNetworkLayerWrapperCallback` simplifies this. The
builder now always creates the base `HttpNetworkLayer`. The provided
callback, if any, receives this `HttpNetworkLayer` and is expected to
return an `HttpTransactionFactory`, typically by wrapping the provided
layer.

This approach:
- Simplifies wrapper logic: Wrappers don't need to create the base
  HttpNetworkLayer themselves.
- Improves testability and clarity: Renamed `http_transaction_factory_`
  to `http_transaction_factory_for_testing_` to make its purpose
  clearer. Added CHECKs to ensure the testing factory and the wrapper
  callback are mutually exclusive.

Change-Id: I836773fa0e5496ec23f62c47a9c00a192a9b7f9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6477687
Reviewed-by: mmenke <mmenke@chromium.org>
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450336}