Fix key storage for App-Bound
StoreKey was allocating a vector of the correct length then inserting
the key and header into this vector. This caused the pref to contain a
bunch of nulls at the end, doubling the pref length.

This did not affect decrypt because extra nulls on the end of encrypted
data are ignored by App-Bound during decryption, but it's still good to
fix the glitch as it's needless storage in prefs, and is functionally
incorrect.

Existing prefs are left, but during a re-encrypt operation they will be
re-written if necessary.

BUG=412730350

Change-Id: I95035872138d8362ca1d0beb4f09a2c3d962eee1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6480463
Reviewed-by: Alex Gough <ajgo@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450347}