LanguageDetector: Cleanup
This cleans up LanguageDetector code a little. Creates a
RunOnDestruction class which generalizes the RejectOnDestructionHelper
so that it can be used in OnModelLoaded to call Cleanup. This ensures
that we don't forget to call it before returning from the function.

Also gets the resolver from AIContextObserver instead of keeping another
copy in LanguageDetectorCreateTask.

Bug: 385173647
Change-Id: I8e56244b586b99bd35e7e584330e3625c9537199
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6476543
Commit-Queue: Nathan Memmott <memmott@chromium.org>
Reviewed-by: Christine Hollingsworth <christinesm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450356}