[iOS][Mim]Allows signout from managed account
The Account Menu requested the SignoutActionSheetCoordinator to
stop. During its stop, the SignoutActionSheetCoordinator call its
completion block, which cause the NTP to reset, causing the account
menu to be stopped, releasing the
SignoutActionSheetCoordinator. However,
SignoutActionSheetCoordinator’s stop is not yet down. As the
SignoutActionSheetCoordinator become a Zombie, the end of its stop
method tries to assign variables of a zombie object and cause the app
to crash.

This change ensures that the SignoutActionSheetCoordinator is retained
until the end of stop. As a side benefit, it ensures that the objects
is stopped at most once.

Bug: 412622695
Change-Id: I912a2459afeb1ecf95e8718f5c61ef63bbce1ffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6479731
Auto-Submit: Arthur Milchior <arthurmilchior@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450425}