[invliadtion] Remove state check in CloudPolicyInvalidator
InvlaidationListener setup for CloudPolicyInvalidator ignores its
internal state and is always observing the listener and ready for
incoming invalidations.

It is possible that the InvalidationListener state changes while core is
disconnected and invalidator is stopped.

Remove the check from OnExpectationChanged and OnInvalidationReceived.
In OnInvalidationReceived, check directly if core is connected, and
ignore an invalidation if disconnected.

Fixed: 412750241
Change-Id: Ia9861036cf2a45135724df0c6df1b172e04d8590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6482470
Commit-Queue: Artem Sumaneev <asumaneev@google.com>
Auto-Submit: Artem Sumaneev <asumaneev@google.com>
Reviewed-by: Victor Gabriel Savu <vsavu@google.com>
Cr-Commit-Position: refs/heads/main@{#1450509}