String constants migration
As a first step for changing base::Environment to take in a
base::cstring_view as parameter, all constants that are eventually
used in this API needs to be fully defined at the call site.

This CL is part of relanding
https://chromium-review.googlesource.com/c/chromium/src/+/6470383.
It has been split into 2 changes to reduce the unnecessary churn if
changing the base::Environment API is reverted again.

This CL was uploaded by git cl split.

R=thomasanderson@chromium.org

Bug: 412679386
Change-Id: Ia00b8c911a898a6a1d6e1290084712f7eef7f221
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6483129
Auto-Submit: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450572}