[PWA] Fix SystemWebApp unit-tests, make InstallApps() private
This CL fixes the SystemWebApp unit-tests to not rely on the
FakeExternallyManagedAppManager and instead use the real version for
testing. This paves the way for making the InstallApps() API private,
further reducing the public API surface of the externally managed app
manager, making it more safe. This is done by:

1. Creating a WebAppInstallInfoFactory that returns a nullptr instead
of a populated info, thus making it easier to test error cases instead
of overriding the install flow.
2. Removing the InstallResultHistogram_ExcludeAlreadyInstalled test, as
SWAs always install from an install info, and can thus never have a
result code of kSuccessAlreadyInstalled.
3. Removing dependency on the SetHandleInstallRequestCallback().

As a result, the InstallApps() API inside the externally managed app
manager has also been made private.

As a side win, this also cleans up any dependencies on the Fake version
of the manager, which will be removed in a future CL.

Fixed: 407549320
Change-Id: Ia9790967d512982478f9917c95ce89134ff488e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6477149
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450574}