Unlink inspector-protocol InputInjector from PaintHolding.
To mitigate a bug with user interactions with a stale page, we are
planning to hold back input events from a page while paint-holding is
active.  Paint-holding is the browser feature to continue to show a
snapshot of the pre-navigation page (instead of a blank page) until
the renderer for the post-navigation page has pushed content to the
GPU.

Because inspector-protocol input injection use-cases are disjoint
from the problem we are fixing in this bug, we are excluding these
injected events from our input-hold-back plan.

Bug: 40057499
Change-Id: I9cf2a569ce6320fa2a870963355e80b495aa1434
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6478308
Commit-Queue: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Vladimir Levin <vmpstr@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450587}