String constants migration
As a first step for changing base::Environment to take in a
base::cstring_view as parameter, all constants that are eventually
used in this API needs to be fully defined at the call site.

This CL is part of relanding
https://chromium-review.googlesource.com/c/chromium/src/+/6470383.
It has been split into 2 changes to reduce the unnecessary churn if
changing the base::Environment API is reverted again.

This CL was uploaded by git cl split.

R=olka@chromium.org

Bug: 412679386
Change-Id: I32bbfc2fc329669b454dc568669a9dd0d3a27ff4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6482848
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Olga Sharonova <olka@chromium.org>
Auto-Submit: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450594}