Fix content stream for chrome/test/data/printing/bug_1399155.pdf
This handcrafted PDF is missing a delimiter before the EI operator in
its content stream. As such, the latest PDFium parsing change cannot
match the EI operator. Add a '\n' before the EI operator to fix this.
Along the way, fix some formatting nits in the PDF as well, and
regenerate it by converting it back to a .in file and running PDFium's
fixup_pdf_template.py on it.

Bug: 407752631
Change-Id: Ib8dc14787443253a0540d371956be2db4e6eb896
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6479449
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1450642}