[UNO-d] Add triggering the sync promo pill for enterprise badging users
It adds triggering the history sync opt-in promo identity pill on
inactivity for the users with enterprise badging enabled. This is a
follow up to crrev.com/c/6467839.

Greetings are not shown for users with enterprise badging. This results
in all history sync opt-in identity pill triggers being 'on inactivity'.

Bug: 407964550
Change-Id: I4d358d08f669b4cce22c31b90eb81417bf36bb10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6482472
Reviewed-by: Amelie Schneider <amelies@google.com>
Commit-Queue: Ernest Nguyen Hung <ernn@google.com>
Reviewed-by: Ryan Sultanem <rsult@google.com>
Cr-Commit-Position: refs/heads/main@{#1450649}